ок

+++ ZooContent/IZooContent.cs
@@ -2,7 +2,7 @@
 {
     public interface IZooContent
     {
-        double Cost { get; set; }
+        float Cost { get; set; }
         string From { get; set; }
         string Name { get; set; }
     }
